requirements.txt
appdirs==1.4.3
certifi==2019.11.28
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
distlib==0.3.0
Django==1.11.29
django-datetimepicker==3.14
docutils==0.16
filelock==3.0.12
gunicorn==19.10.0
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.3.1
Kivy==1.11.1
Kivy-Garden==0.1.4
mysql-connector==2.2.9
mysqlclient==1.3.9
pathlib2==2.3.5
Pygments==2.5.2
PyMySQL==0.9.3
pytz==2019.3
requests==2.23.0
scandir==1.10.0
singledispatch==3.4.0.3
six==1.14.0
typing==3.7.4.1
urllib3==1.25.8
virtualenv==20.0.10
<<<<<<< HEAD
=======
whitenoise==3.3.1
>>>>>>> 6e2698b3eb557ffdea7173accea80e8f369b6223
zipp==1.2.0
